Adeinstien/requirements.txt
Flask
pandas
scikit-learn
numpy
pickle-mixin
gunicorn
annotated-types               0.7.0
anyio                         4.4.0
attrs                         24.2.0
autoscraper                   1.1.14
beautifulsoup4                4.12.3
bs4                           0.0.2
cachetools                    5.4.0
certifi                       2024.7.4
cffi                          1.16.0
charset-normalizer            3.3.2
decorator                     4.4.2
dnspython                     2.6.1
docstring_parser              0.16
elevenlabs                    1.6.1
exceptiongroup                1.2.2
ez_setup                      0.9
filelock                      3.15.4
fsspec                        2024.6.1
google-ai-generativelanguage  0.6.6
google-api-core               2.19.1
google-api-python-client      2.141.0
google-auth                   2.33.0
google-auth-httplib2          0.2.0
google-cloud-aiplatform       1.62.0
google-cloud-bigquery         3.25.0
google-cloud-core             2.4.1
google-cloud-resource-manager 1.12.5
google-cloud-storage          2.18.2
google-crc32c                 1.5.0
google-generativeai           0.7.2
google-resumable-media        2.7.2
google_search_results         2.4.2
googleapis-common-protos      1.63.2
grpc-google-iam-v1            0.13.1
grpcio                        1.65.4
grpcio-status                 1.62.3
h11                           0.14.0
httpcore                      1.0.5
httplib2                      0.22.0
httpx                         0.27.0
huggingface-hub               0.24.5
idna                          3.7
imageio                       2.34.2
imageio-ffmpeg                0.5.1
Jinja2                        3.1.4
lxml                          5.3.0
MarkupSafe                    2.1.5
moviepy                       1.0.3
mpmath                        1.3.0
networkx                      3.2.1
numpy                         2.0.1
oauth2client                  4.1.3
outcome                       1.3.0.post0
packaging                     24.1
pillow                        10.4.0
pip                           24.2
proglog                       0.1.10
proto-plus                    1.24.0
protobuf                      4.25.4
pyasn1                        0.6.0
pyasn1_modules                0.4.0
pycparser                     2.22
pydantic                      2.8.2
pydantic_core                 2.20.1
pymongo                       4.8.0
pyparsing                     3.1.2
PySocks                       1.7.1
python-dateutil               2.9.0.post0
python-dotenv                 1.0.1
PyYAML                        6.0.1
regex                         2024.7.24
requests                      2.32.3
rsa                           4.9
safetensors                   0.4.3
selenium                      4.23.1
setuptools                    70.1.0
shapely                       2.0.5
six                           1.16.0
sniffio                       1.3.1
sortedcontainers              2.4.0
soundfile                     0.12.1
soupsieve                     2.5
sympy                         1.13.1
tokenizers                    0.19.1
torch                         2.4.0
tqdm                          4.66.4
transformers                  4.43.3
trio                          0.26.2
trio-websocket                0.11.1
typing_extensions             4.12.2
uritemplate                   4.1.1
urllib3                       2.2.2
webdriver-manager             4.0.2
websocket-client              1.8.0
websockets                    12.0
wheel                         0.43.0
wsproto                       1.2.0